Fix incorrect ui_hud.dds file (restore sprint indicator)
This commit fixes an error in the previous commit "Revert changes to restore more vanilla behavior". The previously included ui_hud.dds file was incompatible with the latest game patch, causing the sprint indicator to disappear from the HUD.

This commit replaces the incorrect file with the proper ui_hud.dds version, ensuring the sprint indicator is displayed correctly again.